% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradford_analysis.R
\name{bradford_analysis}
\alias{bradford_analysis}
\title{Calculates protein concentration}
\usage{
bradford_analysis(file, sheet_calibration, sheet_sample, dil_fctr, output_dir)
}
\arguments{
\item{file}{path to input xls-file}

\item{sheet_calibration}{sheet containing the standards}

\item{sheet_sample}{sheet containing the samples}

\item{dil_fctr}{numeric value specifying the sample dilution (e.g: 1:10 dilution -> dil_fctr = 10 )}

\item{output_dir}{directory to which the output table and standard curve is saved to as a csv and
pdf file, respectively}
}
\value{
A tibble containing calculated protein concentrations and a standard curve plot
}
\description{
All-in-one function for computing protein concentrations determined by a bradford protein assay.
Requires protein standards of known concentration to plot a standard curve used for the accurate determination
of sample concentrations.

NOTE: Concentration of diluted sample should be in the rang of measured standards.
}
\examples{
bradford_analysis(file = "path/to/Bradford_test.xls", sheet_calibration = 4, sheet_sample = 5,
dil_fctr = 4, output_dir = "/your/favorite/output/directory")
}
